// src/common/enums/swagger.enum.ts
export enum SwaggerApiTags {
    Auth = "Auth",
    User = "User",
    Book = "Book",
    Category = "Category",
    Author = "Author",
    Publisher = "Publisher",
    Basket = "Basket",
    Discount = "Discount",
    Subscription = "Subscription",
}
export enum SwaggerApiConsumes {
    Json = "application/json",
    UrlEncoded = "application/x-www-form-urlencoded",
    FormData = "multipart/form-data",
}